refactor(closed_loop): group generic stages under stages/ subfolder
